xdpsensors.sh: Reduce framerate to 10fps for web image
The web interface only displays about 5 frames or so per second, so
there's no point attempting to create more. Skip 5 frames every
frame in the acquisition, so the framerate is only 10fps. This
makes the load on the system drop below 1 CPU, so it's no longer
occupying a complete core for nothing.